+++ src/main/java/com/treasuredata/android/TreasureData.java
@@ -2,20 +2,15 @@ package com.treasuredata.android;
 
 import android.app.Activity;
 import android.app.Application;
-import android.content.ContentResolver;
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.content.pm.PackageInfo;
-import android.os.AsyncTask;
 import android.os.Build;
 import android.os.Bundle;
-import android.provider.Settings;
 import io.keen.client.java.KeenClient;
 import org.komamitsu.android.util.Log;
 
 import java.io.IOException;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 import java.util.HashMap;
 import java.util.Locale;
 import java.util.Map;
@@ -26,7 +21,7 @@ import java.util.regex.Pattern;
 
 public class TreasureData {
     private static final String TAG = TreasureData.class.getSimpleName();
-    private static final String VERSION = "0.1.17";
+    private static final String VERSION = "0.1.18";
     private static final String LABEL_ADD_EVENT = "addEvent";
     private static final String LABEL_UPLOAD_EVENTS = "uploadEvents";
     private static final Pattern DATABASE_NAME_PATTERN = Pattern.compile("^[0-9a-z_]{3,255}$");
@@ -36,6 +31,9 @@ public class TreasureData {
     private static final String SHARED_PREF_VERSION_KEY = "version";
     private static final String SHARED_PREF_BUILD_KEY = "build";
     private static final String SHARED_PREF_KEY_FIRST_RUN = "first_run";
+    private static final String SHARED_PREF_APP_LIFECYCLE_EVENT_ENABLED = "app_lifecycle_event_enabled";
+    private static final String SHARED_PREF_CUSTOM_EVENT_ENABLED = "custom_event_enabled";
+    private static final String SHARED_PREF_KEY_IS_UNITY = "TDIsUnity";
     private static final String EVENT_KEY_UUID = "td_uuid";
     private static final String EVENT_KEY_SESSION_ID = "td_session_id";
     private static final String EVENT_KEY_SESSION_EVENT = "td_session_event";
@@ -52,17 +50,19 @@ public class TreasureData {
     private static final String EVENT_KEY_PREV_APP_VER_NUM = "td_prev_app_ver_num";
     private static final String EVENT_KEY_LOCALE_COUNTRY = "td_locale_country";
     private static final String EVENT_KEY_LOCALE_LANG = "td_locale_lang";
-    private static final String EVENT_KEY_USER_ID = "td_user_id";
-    private static final String EVENT_KEY_GAID = "td_gaid";
     private static final String EVENT_KEY_EVENT = "td_android_event";
+    private static final String EVENT_KEY_UNITY_EVENT = "td_unity_event";
+    private static final String EVENT_KEY_APP_LIFECYCLE_EVENT_PRIVATE = "__is_app_lifecycle_event";
+    private static final String EVENT_KEY_RESET_UUID_EVENT_PRIVATE = "__is_reset_uuid_event";
     private static final String EVENT_KEY_SERVERSIDE_UPLOAD_TIMESTAMP = "#SSUT";
     private static final String EVENT_DEFAULT_KEY_RECORD_UUID = "record_uuid";
     private static final String OS_TYPE = "Android";
-    private static final String SETTING_ADVERTISING_ID = "advertising_id";
-    private static final String DEFAULT_APP_LIFECYCLE_EVENT_TABLE = "td_app_lifecycle_event";
     private static final String EVENT_APP_INSTALL = "TD_ANDROID_APP_INSTALL";
     private static final String EVENT_APP_OPEN = "TD_ANDROID_APP_OPEN";
     private static final String EVENT_APP_UPDATE = "TD_ANDROID_APP_UPDATE";
+    private static final String EVENT_RESET_UUID = "forget_device_uuid";
+    private static final String TD_DEFAULT_DATABASE = "td";
+    private static final String TD_DEFAULT_TABLE = "td_android";
 
     static {
         TDHttpHandler.VERSION = TreasureData.VERSION;
@@ -73,9 +73,7 @@ public class TreasureData {
 
     private final Context context;
     private final TDClient client;
-    private final String uuid;
-    private volatile String userId;
-    private volatile String gaid;
+    private String uuid;
     private volatile String defaultDatabase;
     private volatile String defaultTable;
     private volatile TDCallback addEventCallBack;
@@ -84,10 +82,11 @@ public class TreasureData {
     private volatile boolean autoAppendModelInformation;
     private volatile boolean autoAppendAppInformation;
     private volatile boolean autoAppendLocaleInformation;
-    private volatile boolean autoTrackApplicationLifecycleEvents = true;
-    private volatile boolean autoTrackApplicationInstalledEvent = true;
-    private volatile boolean autoTrackApplicationOpenEvent = true;
-    private volatile boolean autoTrackApplicationUpdatedEvent = true;
+    private volatile boolean autoTrackAppInstalledEvent = true;
+    private volatile boolean autoTrackAppOpenEvent = true;
+    private volatile boolean autoTrackAppUpdatedEvent = true;
+    private volatile boolean customEventEnabled = true;
+    private volatile boolean appLifecycleEventEnabled;
     private static volatile long sessionTimeoutMilli = Session.DEFAULT_SESSION_PENDING_MILLIS;
     private final String appVersion;
     private final int appVersionNumber;
@@ -100,7 +99,6 @@ public class TreasureData {
         synchronized (TreasureData.class) {
             if (sharedInstance == null) {
                 sharedInstance = new TreasureData(context, apiKey);
-                sharedInstance.initialize();
             }
         }
         return sharedInstance;
@@ -120,15 +118,6 @@ public class TreasureData {
         return sharedInstance;
     }
 
-    private void initialize() {
-        AsyncTask.execute(new Runnable() {
-            @Override
-            public void run() {
-                fetchAdvertisingId();
-            }
-        });
-    }
-
     private SharedPreferences getSharedPreference(Context context) {
         return context.getSharedPreferences(SHARED_PREF_NAME, Context.MODE_PRIVATE);
     }
@@ -145,38 +134,38 @@ public class TreasureData {
         }
     }
 
-    public void setUserId(String userId) {
-        this.userId = userId;
-    }
 
-    private void fetchAdvertisingId() {
-        if("Amazon".equals(Build.MANUFACTURER)) {
-            fetchAmazonAdvertisingId();
-        }else {
-            fetchGoogleAdvertisingId();
+    /**
+     * Reset UUID and send forget_device_uuid event with old uuid
+     */
+    public void resetUniqId() {
+        String targetDatabase = TD_DEFAULT_DATABASE;
+        if (defaultDatabase == null) {
+            Log.w(TAG, "Default database is not set, forget_device_uuid event will be uploaded to " + TD_DEFAULT_DATABASE);
+        } else {
+            targetDatabase = defaultDatabase;
         }
-    }
 
-    private void fetchAmazonAdvertisingId() {
-        ContentResolver cr = context.getContentResolver();
-        gaid = Settings.Secure.getString(cr, SETTING_ADVERTISING_ID);
-    }
+        String targetTable = TD_DEFAULT_TABLE;
+        if (defaultTable == null) {
+            Log.w(TAG, "Default table is not set, forget_device_uuid event will be uploaded to " + TD_DEFAULT_TABLE);
+        } else {
+            targetTable = defaultTable;
+        }
 
-    private void fetchGoogleAdvertisingId() {
-        try {
-            Class AdvertisingIdClient = Class
-                    .forName("com.google.android.gms.ads.identifier.AdvertisingIdClient");
-            Method getAdvertisingInfo = AdvertisingIdClient.getMethod("getAdvertisingIdInfo",
-                    Context.class);
-            Object adInfo = getAdvertisingInfo.invoke(null, context);
-            Method getId = adInfo.getClass().getMethod("getId");
-            gaid = (String) getId.invoke(adInfo);
-        } catch (ClassNotFoundException e) {
-            Log.w(TAG, "Google Play Services SDK not found");
-        } catch (InvocationTargetException e) {
-            Log.w(TAG, "Google Play Services not available");
-        } catch (Exception e) {
-            Log.e(TAG, "Encountered an error connecting to Google Play Services", e);
+        // Send forget_device_uuid event
+        Map record = new HashMap<String, Object>();
+        uuid = getUUID();
+        record.put(isOnUnity() ? EVENT_KEY_UNITY_EVENT : EVENT_KEY_EVENT, EVENT_RESET_UUID);
+        record.put(EVENT_KEY_UUID, uuid);
+        record.put(EVENT_KEY_RESET_UUID_EVENT_PRIVATE, true);
+        addEvent(targetDatabase, targetTable, record);
+
+        // Reset UUID
+        SharedPreferences sharedPreferences = getSharedPreference(context);
+        synchronized (this) {
+            String uuid = UUID.randomUUID().toString();
+            sharedPreferences.edit().putString(SHARED_PREF_KEY_UUID, uuid).commit();
         }
     }
 
@@ -197,7 +186,9 @@ public class TreasureData {
     public TreasureData(Context context, String apiKey) {
         Context applicationContext = context.getApplicationContext();
         this.context = applicationContext;
-        uuid = getUUID();
+        this.uuid = getUUID();
+        this.appLifecycleEventEnabled = getAppLifecycleEventEnabled();
+        this.customEventEnabled = getCustomEventEnabled();
 
         TDClient client = null;
         if (apiKey == null && TDClient.getDefaultApiKey() == null) {
@@ -228,12 +219,12 @@ public class TreasureData {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.ICE_CREAM_SANDWICH) {
             final Application application = (Application) applicationContext;
             application.registerActivityLifecycleCallbacks(new Application.ActivityLifecycleCallbacks() {
-                final AtomicBoolean trackedApplicationLifecycleEvents = new AtomicBoolean(false);
+                final AtomicBoolean trackedAppLifecycleEvents = new AtomicBoolean(false);
 
                 @Override
                 public void onActivityCreated(Activity activity, Bundle bundle) {
-                    if (!trackedApplicationLifecycleEvents.getAndSet(true)
-                            && autoTrackApplicationLifecycleEvents) {
+                    if (!trackedAppLifecycleEvents.getAndSet(true)
+                            && appLifecycleEventEnabled) {
                         trackApplicationLifecycleEvents();
                     }
                 }
@@ -272,12 +263,19 @@ public class TreasureData {
     }
 
     private void trackApplicationLifecycleEvents() {
-        String table = DEFAULT_APP_LIFECYCLE_EVENT_TABLE;
 
+        String targetDatabase = TD_DEFAULT_DATABASE;
+        if (defaultDatabase == null) {
+            Log.w(TAG, "Default database is not set, app lifecycle events will be uploaded to " + TD_DEFAULT_DATABASE);
+        } else {
+            targetDatabase = defaultDatabase;
+        }
+
+        String targetTable = TD_DEFAULT_TABLE;
         if (defaultTable == null) {
-            Log.w(TAG, "Default table is not set, auto tracking events are uploaded to " + DEFAULT_APP_LIFECYCLE_EVENT_TABLE);
-        }else {
-            table = defaultTable;
+            Log.w(TAG, "Default table is not set, auto app lifecycle events will be uploaded to " + TD_DEFAULT_TABLE);
+        } else {
+            targetTable = defaultTable;
         }
 
         String currentVersion = appVersion;
@@ -288,28 +286,31 @@ public class TreasureData {
         int previousBuild = sharedPreferences.getInt(SHARED_PREF_BUILD_KEY, 0);
 
         Map<String, Object> record;
-        if (autoTrackApplicationInstalledEvent && previousBuild == 0) {
+        if (autoTrackAppInstalledEvent && previousBuild == 0) {
             record = new HashMap<String, Object>();
             record.put(EVENT_KEY_EVENT, EVENT_APP_INSTALL);
             record.put(EVENT_KEY_APP_VER_NUM, currentBuild);
             record.put(EVENT_KEY_APP_VER, currentVersion);
-            addEvent(table, record);
-        }else if (autoTrackApplicationUpdatedEvent && currentBuild != previousBuild) {
+            record.put(EVENT_KEY_APP_LIFECYCLE_EVENT_PRIVATE, true);
+            addEvent(targetDatabase, targetTable, record);
+        }else if (autoTrackAppUpdatedEvent && currentBuild != previousBuild) {
             record = new HashMap<String, Object>();
             record.put(EVENT_KEY_EVENT, EVENT_APP_UPDATE);
             record.put(EVENT_KEY_APP_VER_NUM, currentBuild);
             record.put(EVENT_KEY_APP_VER, currentVersion);
             record.put(EVENT_KEY_PREV_APP_VER_NUM, previousBuild);
             record.put(EVENT_KEY_PREV_APP_VER, previousVersion);
-            addEvent(table, record);
+            record.put(EVENT_KEY_APP_LIFECYCLE_EVENT_PRIVATE, true);
+            addEvent(targetDatabase, targetTable, record);
         }
 
-        if (autoTrackApplicationOpenEvent) {
+        if (autoTrackAppOpenEvent) {
             record = new HashMap<String, Object>();
             record.put(EVENT_KEY_EVENT, EVENT_APP_OPEN);
             record.put(EVENT_KEY_APP_VER_NUM, currentBuild);
             record.put(EVENT_KEY_APP_VER, currentVersion);
-            addEvent(table, record);
+            record.put(EVENT_KEY_APP_LIFECYCLE_EVENT_PRIVATE, true);
+            addEvent(targetDatabase, targetTable, record);
         }
 
         SharedPreferences.Editor editor = sharedPreferences.edit();
@@ -396,6 +397,19 @@ public class TreasureData {
     }
 
     public void addEventWithCallback(String database, String table, Map<String, Object> origRecord, TDCallback callback) {
+
+        if(!isCustomEventEnabled() && isCustomEvent(origRecord)) {
+            return;
+        }
+
+        if(!isAppLifecycleEventEnabled() && isAppLifecycleEvent(origRecord)) {
+            return;
+        }
+
+        // Remove private key
+        origRecord.remove(EVENT_KEY_APP_LIFECYCLE_EVENT_PRIVATE);
+        origRecord.remove(EVENT_KEY_RESET_UUID_EVENT_PRIVATE);
+
         if (client == null) {
             Log.w(TAG, "TDClient is null");
             return;
@@ -418,14 +432,6 @@ public class TreasureData {
         Map<String, Object> record = new HashMap<String, Object>();
         if (origRecord != null) {
             record.putAll(origRecord);
-        }
-
-        if (gaid != null) {
-            record.put(EVENT_KEY_GAID, gaid);
-        }
-
-        if (userId != null) {
-            record.put(EVENT_KEY_USER_ID, userId);
         }
 
         appendSessionId(record);
@@ -484,6 +490,7 @@ public class TreasureData {
     }
 
     public void uploadEventsWithCallback(TDCallback callback) {
+
         if (client == null) {
             Log.w(TAG, "TDClient is null");
             return;
@@ -579,20 +586,101 @@ public class TreasureData {
         record.put(autoAppendRecordUUIDColumn, UUID.randomUUID().toString());
     }
 
-    public void disableTrackApplicationLifecycleEvents() {
-        this.autoTrackApplicationLifecycleEvents = false;
+    /**
+     * Enable app lifecycle tracking. This setting has no effect to custom tracking
+     */
+    public void enableAppLifecycleEvent() {
+        enableAppLifecycleEvent(true);
     }
 
-    public void disableApplicationInstalledEvent() {
-        this.autoTrackApplicationInstalledEvent = false;
+    /**
+     * Disable app lifecycle tracking. This setting has no effect to custom tracking
+     */
+    public void disableAppLifecycleEvent() {
+        enableAppLifecycleEvent(false);
     }
 
-    public void disableApplicationUpdatedEvent() {
-        this.autoTrackApplicationUpdatedEvent = false;
+
+    public void enableAppLifecycleEvent(boolean enabled) {
+        this.appLifecycleEventEnabled = enabled;
+        SharedPreferences sharedPreferences = getSharedPreference(context);
+        synchronized (this) {
+            sharedPreferences.edit().putBoolean(SHARED_PREF_APP_LIFECYCLE_EVENT_ENABLED,  this.appLifecycleEventEnabled).commit();
+        }
     }
 
-    public void disableApplicationOpenEvent() {
-        this.autoTrackApplicationOpenEvent = false;
+    /**
+     * Whether or not the app lifecycle tracking is enabled
+     * @return true : enabled, false : disabled
+     */
+    public boolean isAppLifecycleEventEnabled() {
+        return this.appLifecycleEventEnabled;
+    }
+
+    private boolean getAppLifecycleEventEnabled() {
+        SharedPreferences sharedPreferences = getSharedPreference(context);
+        synchronized (this) {
+            return sharedPreferences.getBoolean(SHARED_PREF_APP_LIFECYCLE_EVENT_ENABLED, false);
+        }
+    }
+
+
+    /**
+     * Enable custom event tracking. This setting has no effect to auto tracking
+     */
+    public void enableCustomEvent() {
+        enableCustomEvent(true);
+    }
+
+    /**
+     * Disable custom event tracking. This setting has no effect to auto tracking
+     */
+    public void disableCustomEvent() {
+        enableCustomEvent(false);
+    }
+
+    /**
+     * Whether or not the custom event tracking is enable
+     * @return true : enable, false : disabled
+     */
+    public boolean isCustomEventEnabled() {
+        return this.customEventEnabled;
+    }
+
+    public void enableCustomEvent(boolean enabled) {
+        this.customEventEnabled = enabled;
+        SharedPreferences sharedPreferences = getSharedPreference(context);
+        synchronized (this) {
+            sharedPreferences.edit().putBoolean(SHARED_PREF_CUSTOM_EVENT_ENABLED,  this.customEventEnabled).commit();
+        }
+    }
+
+    private boolean getCustomEventEnabled() {
+        SharedPreferences sharedPreferences = getSharedPreference(context);
+        synchronized (this) {
+            this.customEventEnabled = sharedPreferences.getBoolean(SHARED_PREF_CUSTOM_EVENT_ENABLED, true);
+            return this.customEventEnabled;
+        }
+    }
+
+    private boolean isCustomEvent(Map record) {
+        return !record.containsKey(EVENT_KEY_APP_LIFECYCLE_EVENT_PRIVATE) && !record.containsKey(EVENT_KEY_RESET_UUID_EVENT_PRIVATE);
+    }
+
+    private boolean isAppLifecycleEvent(Map record) {
+        return record.containsKey(EVENT_KEY_APP_LIFECYCLE_EVENT_PRIVATE);
+    }
+
+    public void disableAppInstalledEvent() {
+        this.autoTrackAppInstalledEvent = false;
+    }
+
+    public void disableAppUpdatedEvent() {
+        this.autoTrackAppUpdatedEvent = false;
+    }
+
+    public void disableAppOpenEvent() {
+        this.autoTrackAppOpenEvent = false;
     }
 
     public void disableAutoAppendUniqId() {
@@ -733,6 +821,14 @@ public class TreasureData {
 
     public void disableAutoAppendRecordUUID() {
         autoAppendRecordUUIDColumn = null;
+    }
+
+    private boolean isOnUnity() {
+        return context
+                // Default preference name of Unity's PlayerPrefs
+                .getSharedPreferences(context.getPackageName() + ".v2.playerprefs", Context.MODE_PRIVATE)
+                // Since Unity's PlayerPrefs doesn't support boolean
+                .getInt(SHARED_PREF_KEY_IS_UNITY, 0) == 1;
     }
 
     // Only for testing